% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache_dir}
\alias{cache_dir}
\title{Get the cache directory path}
\usage{
cache_dir()
}
\value{
Path string
}
\description{
Defaults to ~/.wheather/cache, configurable via
options(wheather.cache_dir = "...").
}
