% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{date_seq}
\alias{date_seq}
\title{Generate a sequence of dates}
\usage{
date_seq(start, end)
}
\arguments{
\item{start}{Start date (Date or character YYYY-MM-DD)}

\item{end}{End date (Date or character YYYY-MM-DD)}
}
\value{
Character vector of dates in YYYY-MM-DD format
}
\description{
Generate a sequence of dates
}
