% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{score_period}
\alias{score_period}
\title{Score all days in a dataset}
\usage{
score_period(data, weights = default_weights(), params = default_params())
}
\arguments{
\item{data}{A data.table from \code{\link[=fetch_date_range]{fetch_date_range()}}}

\item{weights}{Scoring weights}

\item{params}{Scoring parameters}
}
\value{
The input data.table with added score columns
}
\description{
Adds score columns to a weather data.table.
}
