% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{top_cities}
\alias{top_cities}
\title{Get top cities by population}
\usage{
top_cities(n = 1000)
}
\arguments{
\item{n}{Number of cities to return (default 1000)}
}
\value{
A data.table with columns: name, country, lat, lon, pop, capital
}
\description{
Returns a data.table of the most populated cities worldwide,
sourced from the \code{maps} package's world.cities dataset.
}
