% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{save_to_cache}
\alias{save_to_cache}
\title{Save weather data to the local cache}
\usage{
save_to_cache(data, lat, lon, date)
}
\arguments{
\item{data}{A data.table (single row) of daily weather data}

\item{lat}{Latitude}

\item{lon}{Longitude}

\item{date}{Date string (YYYY-MM-DD)}
}
\description{
Save weather data to the local cache
}
